Reset the state of matrix object using a proper reset method.
The Matrix method 'set(null)', with null parameter, works fine on emulators and real
devices, but doesn't play well in unit tests with Robolectric, which
throw an Exception after checking for null params.

Since a proper method for reseting the state of Matrixes are available,
should not be an issue using it.